block要素を表現するjsonをBlockPayload下のファイルに移動
